+++ angular_router/CHANGELOG.md
@@ -1,15 +1,21 @@
+- Fixed a bug where `RouterLinkDirective` was not keyboard accessible.
+
 ## 2.0.0-alpha+1
+
 - Support for angular 5.0.0-alpha+1
 
 ## 2.0.0-alpha
+
 - Major refactoring of the `angular_router` package. For a migration guide from
 `angular_router` v1, see
 https://github.com/dart-lang/angular/blob/master/angular_router/g3doc/migration_guide.md.
 
 ## 1.0.2
+
 - Support for angular 4.0.0.
 
 ## 1.0.1
+
 - Minor internal changes to support angular 4.0.0-beta
 
 ## 1.0.0